requirements.txt
numpy~=1.26.4
ipdb~=0.13.13
opencv-python~=4.11.0.86
#torch~=2.7.1+cu128
#torchvision~=0.22.1+cu128
matplotlib
Pillow~=10.1.0
tqdm~=4.67.1
ultralytics~=8.3.153
imageio~=2.37.0
ultralytics-thop~=2.0.14
requests~=2.32.3
psutil~=7.0.0
PyYAML~=6.0.2
scipy
torchsummary~=1.5.1
tensorboardX~=2.6.4
tensorflow~=2.15.0
pycocotools~=2.0.10
packaging~=25.0
torchgeometry~=0.1.2
ImageAug~=0.1.0.post0
pandas~=2.3.0
ipython
comm~=0.2.2

onnx~=1.18.0
onnx2torch~=1.5.15